Handle results loading errors and refine tests.

+++ front-end/src/pages/Evaluator/utils/evaluatorSearchSchema.ts
@@ -7,7 +7,7 @@ import minMaxValidator from './minMaxValidator'
 
 export const evaluatorSchema = z.object({
   view: fallback(z.enum(['all', 'sample']), 'sample'),
-  page: fallback(z.number().gt(0), 1),
+  page: fallback(z.number().gt(0).int(), 1),
   page_size: fallback(z.number().gt(0), ITEMS_PER_PAGE),
   amt_past_due_min: minMaxValidator,
   amt_past_due_max: minMaxValidator,
